+++ CHANGES.md
@@ -5,6 +5,7 @@ Next version (SNAPSHOT)
 - Build system updated to Gradle.
 - cleaned Python PEP 8 lints for jnc.py
 - reduced Java lints (PMD tool, "java" quickstart.xml lint set)
+- implemented NETCONF-1.1 framing
 
 2012-10-17
 - JNC is finally made open source! There are some changes remaining